Eliminate Aequitas::Context.
Context was a holdover from DM. Its purpose is to allow related instances to attempt validation in the same context as the root instance is validated when validating through a graph of objects.

This global was used because DM did not explicitly pass the root object's validation context to related instances as they were validated (in DM the save (& therefore #valid?) call cascades via #save_parents & #save_children).

Since Aequitas has no concept of validating a graph of objects (at this point), this feature is not needed. If, at some point in the future, Aequitas needs to be able to validate a graph of objects, we can devise a better implementation at that time.